@@ -1,16 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<OutputType>Library</OutputType>
-		<IsPackable>false</IsPackable>
-		<IsTestProject>true</IsTestProject>
-		<Product>Unit Tests</Product>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
-		<PackageReference Include="xunit" Version="2.7.0" />
+	    <ProjectReference Include="..\src\AssemblyLoader.csproj" />
 	</ItemGroup>
 
 </Project>
